Extract dropdown to countries dropdown
