+++ src/App.js
@@ -1,4 +1,4 @@
-import CountriesDropdown from "./components/countriesDropdown";
+import CountriesDropdown from "./components/common/countriesDropdown";
 
 import "./App.css";
